fix(consts): uid_t-typed host id accessors, neutral diagnostic event
resolveHostID now parses CLAWKER_HOST_UID/GID via ParseUint(_,10,32) so
over-uint32 inputs become a structured "malformed" fallback instead of
silently wrapping when userStage casts to uint32 (closes CodeQL #285,
#286). HostUID()/HostGID() and HostIDResolution.Value return uint32, so
the PipeStage assignment is a total identity at the call site.

Diagnostic event renamed host_uid_unavailable -> host_id_unavailable;
the env field on the record (CLAWKER_HOST_UID or CLAWKER_HOST_GID)
already disambiguates UID vs GID for the operator.

Also collapses the duplicated logHostIdentity doc paragraph and
qualifies the bare HostUID() reference in workspace/setup.go.

+++ internal/controlplane/agent/init.go
@@ -699,8 +699,8 @@ func (e *Executor) plan() []step {
 func userStage(script string) *clawkerdv1.PipeStage {
 	return &clawkerdv1.PipeStage{
 		Argv: []string{"sh", "-c", script},
-		Uid:  uint32(consts.HostUID()),
-		Gid:  uint32(consts.HostGID()),
+		Uid:  consts.HostUID(),
+		Gid:  consts.HostGID(),
 		Env: map[string]string{
 			"HOME": consts.ContainerHomeDir,
 			"USER": consts.ContainerUser,
